+++ lib/cocos2d-x/cocos2dx/tilemap_parallax_nodes/CCTMXLayer.cpp
@@ -390,7 +390,7 @@ CCSprite * CCTMXLayer::insertTileForGID(unsigned int gid, const CCPoint& pos)
     setupTileSprite(tile, pos, gid);
 
     // get atlas index
-    unsigned int indexForZ = atlasIndexForNewZ(z);
+    unsigned int indexForZ = atlasIndexForNewZ((int)z);
 
     // Optimization: add the quad without adding a child
     this->insertQuadFromSprite(tile, indexForZ);
@@ -477,7 +477,7 @@ unsigned int CCTMXLayer::atlasIndexForExistantZ(unsigned int z)
 
     CCAssert(item, "TMX atlas index not found. Shall not happen");
 
-    int index = ((size_t)item - (size_t)m_pAtlasIndexArray->arr) / sizeof(void*);
+    int index = (int)((size_t)item - (size_t)m_pAtlasIndexArray->arr) / sizeof(void*);
     return index;
 }
 unsigned int CCTMXLayer::atlasIndexForNewZ(int z)
@@ -486,7 +486,7 @@ unsigned int CCTMXLayer::atlasIndexForNewZ(int z)
     unsigned int i=0;
     for (i=0; i< m_pAtlasIndexArray->num ; i++) 
     {
-        int val = (size_t) m_pAtlasIndexArray->arr[i];
+        size_t val = (size_t) m_pAtlasIndexArray->arr[i];
         if (z < val)
         {
             break;
@@ -568,7 +568,7 @@ void CCTMXLayer::removeChild(CCNode* node, bool cleanup)
     CCAssert(m_pChildren->containsObject(sprite), "Tile does not belong to TMXLayer");
 
     unsigned int atlasIndex = sprite->getAtlasIndex();
-    unsigned int zz = (size_t)m_pAtlasIndexArray->arr[atlasIndex];
+    unsigned int zz = (unsigned int)((size_t)m_pAtlasIndexArray->arr[atlasIndex]);
     m_pTiles[zz] = 0;
     ccCArrayRemoveValueAtIndex(m_pAtlasIndexArray, atlasIndex);
     CCSpriteBatchNode::removeChild(sprite, cleanup);
